Fix (in)direct positionning
